+++ packages/app/src/report/components/detection/Detection.tsx
@@ -22,5 +22,6 @@ const useStyles = makeStyles()(() => ({
   detection: {
     height: "100%",
     position: "relative",
+    overflow: "hidden",
   },
 }));