+++ packages/backend/src/megalinter/parsers/parse-megalinter-start.ts
@@ -1,5 +1,6 @@
 import { LinterStatus } from "shared-types";
 import { ReportStore } from "../../stores/be-report-store";
+import { logger } from "../../utils/logger";
 import {
   MegalinterStartMessage,
   MessageType,
@@ -11,13 +12,21 @@ export const parseMegalinterStart = (
   reportStore: ReportStore
 ) => {
   if (msg.messageType === MessageType.MegalinterStart) {
-    const linters = msg.linters.filter(filterLinters).map((linter) => ({
-      name: linter.linterId,
-      errors: 0,
-      severity: undefined,
-      docUrl: linter.docUrl,
-      status: linter.linterStatus || LinterStatus.Started,
-    }));
+    const linters = msg.linters.filter(filterLinters).map((linter) => {
+      if (!linter.iconPngUrl) {
+        logger.megalinter.error(
+          `Missing "iconPngUrl" for:  ${linter.linterId}`
+        );
+      }
+      return {
+        name: linter.linterId,
+        errors: 0,
+        severity: undefined,
+        docUrl: linter.docUrl,
+        status: linter.linterStatus || LinterStatus.Started,
+        logoUrl: linter.iconPngUrl,
+      };
+    });
 
     reportStore.set({ linters });
   }